Simple test to show FSM minimal functionality + BLE reading from the client

+++ sl_bt_m20_strain_sensor.c
@@ -11,25 +11,25 @@ sl_status_t sl_bt_torque_send_data(uint8_t* number, size_t size){
   sl_status_t sc;
   sc = sl_bt_gatt_server_write_attribute_value(gattdb_strain_1_characteristic, 0, size/4, number);
   if (sc == SL_STATUS_OK) {
-      app_log_info("Attribute written: 0x%f", (float)*number);
+      app_log_info("Attribute written: 0x%hu", (uint16_t)*number);
     }
 
   number += size/4;
   sc = sl_bt_gatt_server_write_attribute_value(gattdb_strain_2_characteristic, 0, size/4, number);
     if (sc == SL_STATUS_OK) {
-        app_log_info("Attribute written: 0x%f", (float)*number);
+        app_log_info("Attribute written: 0x%hu", (uint16_t)*number);
       }
 
   number += size/4;
   sc = sl_bt_gatt_server_write_attribute_value(gattdb_strain_3_characteristic, 0, size/4, number);
     if (sc == SL_STATUS_OK) {
-        app_log_info("Attribute written: 0x%f", (float)*number);
+        app_log_info("Attribute written: 0x%hu", (uint16_t)*number);
       }
 
   number += size/4;
   sc = sl_bt_gatt_server_write_attribute_value(gattdb_temp_1_characteristic, 0, size/4, number);
     if (sc == SL_STATUS_OK) {
-        app_log_info("Attribute written: 0x%f", (float)*number);
+        app_log_info("Attribute written: 0x%hu", (uint16_t)*number);
       }
   return sc;
 }
